Populates Visit_id differently, adds msg builder
NICOLETTE! Changed the way we populate visit_id, using a value from the
patient segment. Got this from talking with Marco after class. WE WILL
HAVE TO SEND THIS VISIT ID WITH THE ORC|SN

Also, there is a new mirth channel for building messages, it’s Marco’s
example, but we might be able build one similar.